fix api and simplify how it's works

+++ DiabloII.Items.Api/Items/Models/ItemProperty.cs
@@ -12,6 +12,6 @@ namespace DiabloII.Items.Api.Items.Responses
         public int Maximum { get; set; }
         public bool IsPercent { get; set; }
 		public string FirstChararacter { get; set; }
-		public int OrderIndex { get; set; }
+		public double OrderIndex { get; set; }
 	}
 }
